requirements.txt
spotipy>=2.23.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
colorthief>=0.2.1
flask>=3.0.0
flask-cors>=4.0.0
waitress>=2.1.2
gunicorn>=21.2.0
